Add TimePicker toolbar and fully replace old TimePickerRoot
