+++ lib/src/DateTimePicker/DateTimePickerRoot.tsx
@@ -11,7 +11,7 @@ import { useUtils } from '../_shared/hooks/useUtils';
 import { MaterialUiPickersDate } from '../typings/date';
 import { convertToMeridiem } from '../_helpers/time-utils';
 import { BaseDatePickerProps } from '../DatePicker/DatePickerRoot';
-import { BaseTimePickerProps, useMeridiemMode } from '../TimePicker/TimePickerRoot';
+import { BaseTimePickerProps, useMeridiemMode } from '../TimePicker/TimePickerToolbar';
 
 export type DateTimePickerViewType = 'year' | 'date' | 'hours' | 'minutes';
